backend/requirements.txt
fastapi
uvicorn
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy
psycopg2-binary
redis
celery
python-jose[cryptography]
passlib
bcrypt
python-multipart
python-dotenv
requests
openai
google-generativeai
pymupdf
nltk
spacy
sentence-transformers
chromadb>=0.5.0
langchain
pytest
httpx
alembic
